// frontend/src/gen/api/src/models/CommentsListDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NewLevel
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CommentsListDto
 */
export interface CommentsListDto {
    /**
     * 
     * @type {string}
     * @memberof CommentsListDto
     */
    comment?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CommentsListDto
     */
    userName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CommentsListDto
     */
    userAvatarSrc?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof CommentsListDto
     */
    dateOfComment?: Date;
}

/**
 * Check if a given object implements the CommentsListDto interface.
 */
export function instanceOfCommentsListDto(value: object): value is CommentsListDto {
    return true;
}

export function CommentsListDtoFromJSON(json: any): CommentsListDto {
    return CommentsListDtoFromJSONTyped(json, false);
}

export function CommentsListDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CommentsListDto {
    if (json == null) {
        return json;
    }
    return {
        
        'comment': json['comment'] == null ? undefined : json['comment'],
        'userName': json['userName'] == null ? undefined : json['userName'],
        'userAvatarSrc': json['userAvatarSrc'] == null ? undefined : json['userAvatarSrc'],
        'dateOfComment': json['dateOfComment'] == null ? undefined : (new Date(json['dateOfComment'])),
    };
}

export function CommentsListDtoToJSON(json: any): CommentsListDto {
    return CommentsListDtoToJSONTyped(json, false);
}

export function CommentsListDtoToJSONTyped(value?: CommentsListDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'comment': value['comment'],
        'userName': value['userName'],
        'userAvatarSrc': value['userAvatarSrc'],
        'dateOfComment': value['dateOfComment'] == null ? undefined : ((value['dateOfComment']).toISOString()),
    };
}