// frontend/src/gen/api/src/models/ETargetType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NewLevel
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ETargetType = {
    NUMBER_0: 0,
    NUMBER_1: 1,
    NUMBER_2: 2,
    NUMBER_3: 3
} as const;
export type ETargetType = typeof ETargetType[keyof typeof ETargetType];


export function instanceOfETargetType(value: any): boolean {
    for (const key in ETargetType) {
        if (Object.prototype.hasOwnProperty.call(ETargetType, key)) {
            if (ETargetType[key as keyof typeof ETargetType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ETargetTypeFromJSON(json: any): ETargetType {
    return ETargetTypeFromJSONTyped(json, false);
}

export function ETargetTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): ETargetType {
    return json as ETargetType;
}

export function ETargetTypeToJSON(value?: ETargetType | null): any {
    return value as any;
}

export function ETargetTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): ETargetType {
    return value as ETargetType;
}